modified:   src/main/webapp/AddBook.jsp 	modified:   src/main/webapp/bookstatus.jsp 	modified:   src/main/webapp/delete.jsp 	modified:   src/main/webapp/viewuser.jsp

+++ src/main/webapp/delete.jsp
@@ -3,25 +3,7 @@
 out.println("<h1 align=center>Welcome "+session.getAttribute("username")+"</h1>");
 out.println("<a align=center href=https://tomcat-sample.herokuapp.com/home.jsp><h1>HOME</h1></a>");
 %>
-<%!
-		String name,query;
-		Connection con;
-		Statement st;
-		int res;
-	%>
-	<%
-		try{
-			String name=(String)session.getAttribute("username");
-			Class.forName("org.postgresql.Driver");
-			 String dbUrl = System.getenv("JDBC_DATABASE_URL");
-			con=DriverManager.getConnection(dbUrl);
-			st=con.createStatement();
-		  }
-			catch(Exception e){
-				out.println(e);
-			}
-	%>	
-		<html>
+<html>
 		<head>
 			<link rel="stylesheet" type="text/css" href="style.css">
 			<link href="https://fonts.googleapis.com/css?family=Josefin+Sans" rel="stylesheet">
@@ -38,16 +20,38 @@ out.println("<a align=center href=https://tomcat-sample.herokuapp.com/home.jsp><
 			</fieldset>
 			</div>
 		</body>
-		</html>
+</html>
+<%!
+		String name,query;
+		Connection con;
+		Statement st;
+		int res;
+	%>
+	<%
+		try{
+			String name=(String)session.getAttribute("username");
+			Class.forName("org.postgresql.Driver");
+			 String dbUrl = System.getenv("JDBC_DATABASE_URL");
+			con=DriverManager.getConnection(dbUrl);
+			st=con.createStatement();
+			name=request.getParameter("book");
+			if(name!=null){
+			query="delete from Book where BookId='"+name+"'";
+			res=st.executeUpdate(query);
+			}
+			if(res>0)
+				out.println("Deleted Successfully");
+		  }
+			catch(Exception e){
+				out.println(e);
+			}
+	%>	
+		
 		
 		
 	<%
 		try{
-			name=request.getParameter("book");
-			query="delete from Book where BookId='"+name+"'";
-			res=st.executeUpdate(query);
-			if(res>0)
-				out.println("Deleted Successfully");
+			
 		  }
 			catch(Exception e){
 				out.println(e);
